Refactor story
